backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Cors==4.0.0
Flask-Admin==1.6.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.27
boto3==1.34.69 